feat(Figma): Added comments, updated mapper. Removed unused data models
