+++ lib/data/mapper/figma/figma_nodes_mapper.dart
@@ -1,5 +1,5 @@
 import 'package:onix_flutter_bricks/core/arch/domain/common/converter/mapper.dart';
-import 'package:onix_flutter_bricks/data/mapper/figma/paint_property_mapper.dart';
+import 'package:onix_flutter_bricks/data/mapper/figma/properties_mapper.dart';
 import 'package:onix_flutter_bricks/data/model/figma/nodes/connector/connector_node_data_model.dart';
 import 'package:onix_flutter_bricks/data/model/figma/nodes/document/document_node_data_model.dart';
 import 'package:onix_flutter_bricks/data/model/figma/nodes/frame/frame_node_data_model.dart';
@@ -11,6 +11,7 @@ import 'package:onix_flutter_bricks/data/model/figma/nodes/table_cell/table_cell
 import 'package:onix_flutter_bricks/data/model/figma/nodes/text/text_node_data_model.dart';
 import 'package:onix_flutter_bricks/data/model/figma/nodes/vector/vector_node_data_model.dart';
 import 'package:onix_flutter_bricks/data/model/figma/properties/paint/paint_property_data_model.dart';
+import 'package:onix_flutter_bricks/data/model/figma/properties/type_style/figma_type_style.dart';
 import 'package:onix_flutter_bricks/domain/entity/figma/nodes/connector_node/connector_node_entity.dart';
 import 'package:onix_flutter_bricks/domain/entity/figma/nodes/document_node/document_node_entity.dart';
 import 'package:onix_flutter_bricks/domain/entity/figma/nodes/frame_node/frame_node_entity.dart';
@@ -23,9 +24,6 @@ import 'package:onix_flutter_bricks/domain/entity/figma/nodes/table_cell_node/ta
 import 'package:onix_flutter_bricks/domain/entity/figma/nodes/text_node/text_node_entity.dart';
 import 'package:onix_flutter_bricks/domain/entity/figma/nodes/vector_node/vector_node_entity.dart';
 import 'package:onix_flutter_bricks/domain/entity/figma/properties/paint_property/paint_property.dart';
-import 'package:onix_flutter_bricks/domain/entity/figma/properties/type_style/type_style_metadata.dart';
-
-typedef A = ({int a, int b});
 
 class FigmaNodesMapper {
   List<BaseNode> mapNodesDataModelToEntity(List<NodeDataModel> from) =>
@@ -43,6 +41,7 @@ class _MapNodeDataModelToEntity
           key: from.key ?? '',
           name: from.name ?? '',
           type: from.type ?? '',
+          // (Ivan Modlo): Maybe we should call the map method recursively
           children: from.children
                   ?.map(
                     (e) => BaseNodeEntity(
@@ -62,11 +61,14 @@ class _MapNodeDataModelToEntity
           name: from.name ?? '',
           type: from.type ?? '',
           fills: _mapPaintProperty(from.fills),
-          style: TypeStyleMetadata(
-            fontFamily: from.style?.fontFamily ?? '',
-            fontWeight: from.style?.fontWeight ?? 0,
-            fontSize: from.style?.fontSize ?? 0,
-            letterSpacing: from.style?.letterSpacing ?? 0,
+          style: PropertyMapper().mapFigmaTypeStyleDataModelToEntity(
+            from.style ??
+                const FigmaTypeStyle(
+                  fontFamily: '',
+                  fontWeight: 0,
+                  fontSize: 0,
+                  letterSpacing: 0,
+                ),
           ),
         );
       case FrameNodeDataModel _:
@@ -138,7 +140,5 @@ class _MapNodeDataModelToEntity
   }
 
   List<PaintProperty> _mapPaintProperty(List<PaintPropertyDataModel>? from) =>
-      PaintPropertyMapper()
-          .mapPaintPropertyDataModelToEntity(from ?? [])
-          .toList();
+      PropertyMapper().mapPaintPropertyDataModelToEntity(from ?? []).toList();
 }
