+++ lib/data/model/figma/nodes/frame/frame_node_data_model.dart
@@ -4,6 +4,7 @@ import 'package:onix_flutter_bricks/data/model/figma/properties/paint/paint_prop
 
 part 'frame_node_data_model.g.dart';
 
+// https://www.figma.com/developers/api#frame-props
 @JsonSerializable()
 class FrameNodeDataModel extends FillPaintsNodeDataModel {
   const FrameNodeDataModel({
