+++ lib/data/model/figma/responses/node/figma_nodes_response.dart
@@ -1,5 +1,5 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
-import 'package:onix_flutter_bricks/data/mapper/converter/figma/figma_nodes_converter.dart';
+import 'package:onix_flutter_bricks/data/converter/figma/figma_nodes_converter.dart';
 import 'package:onix_flutter_bricks/data/model/figma/nodes/node/node_data_model.dart';
 
 part 'figma_nodes_response.g.dart';
