+++ lib/domain/entity/figma/nodes/node/base_node.dart
@@ -1,3 +1,4 @@
+// https://www.figma.com/developers/api#node-type
 abstract class BaseNode {
   String get id;
   String get key;
