+++ lib/domain/entity/figma/nodes/shape_with_text_node/shape_with_text_node_entity.dart
@@ -4,6 +4,7 @@ import 'package:onix_flutter_bricks/domain/entity/figma/properties/paint_propert
 
 part 'shape_with_text_node_entity.freezed.dart';
 
+// https://www.figma.com/developers/api#shape_with_text-props
 @freezed
 class ShapeWithTextNodeEntity
     with _$ShapeWithTextNodeEntity
