+++ lib/domain/entity/figma/properties/color_property/color_property.dart
@@ -3,7 +3,6 @@ import 'package:freezed_annotation/freezed_annotation.dart';
 part 'color_property.freezed.dart';
 
 // https://www.figma.com/developers/api#color-type
-
 @freezed
 class ColorProperty with _$ColorProperty {
   const factory ColorProperty({
