+++ lib/domain/entity/figma/properties/type_style/type_style_metadata.dart
@@ -3,7 +3,6 @@ import 'package:freezed_annotation/freezed_annotation.dart';
 part 'type_style_metadata.freezed.dart';
 
 // https://www.figma.com/developers/api#style-type
-
 @freezed
 class TypeStyleMetadata with _$TypeStyleMetadata {
   const factory TypeStyleMetadata({
